// server/model/album.ts
import {
  Association, 
  DataTypes, 
  Model, 
  Sequelize, 
  HasManyCountAssociationsMixin, 
  HasManyGetAssociationsMixin, 
  HasManyHasAssociationMixin, 
  BelongsToSetAssociationMixin, 
  BelongsToGetAssociationMixin
} from 'sequelize';

import { Artist } from './artist';
import { Token } from './token';

class Album extends Model {
  public name!: string;
  public year!: Date;
  public description?: string;

  //Auto-generated
  public id!: number;
  public createdAt!: Date;
  public updatedAt!: Date;

  //Artist association methods

  //Token association methods
  
  // Populated for inclusions
  public readonly artist!: Artist;
  public readonly tokens?: Token[];

  public static associations: {
    artist: Association<Artist, Album>;
    tokens: Association<Album, Token>;
  }

  public static initialize(sequelize: Sequelize) {
    this.init(
      {
        name: {
          type: DataTypes.STRING,
          allowNull: false
        },
        year: {
          type: DataTypes.DATE,
          allowNull: false
        },
        description: {
          type: DataTypes.STRING,
        }
      }, 
    {sequelize})
  }
}
export { Album };